Move domain-search queries to api-queries package
