+++ packages/api-queries/src/domain-availability.ts
@@ -5,8 +5,5 @@ export const domainAvailabilityQuery = ( domainName: string ) =>
 	queryOptions( {
 		queryKey: [ 'domain-availability', domainName ],
 		queryFn: () => fetchDomainAvailability( domainName ),
-		refetchOnWindowFocus: false,
-		refetchOnMount: false,
-		enabled: false,
 		meta: { persist: false },
 	} );
