+++ packages/api-queries/src/domains.ts
@@ -1,4 +1,8 @@
-import { fetchDomainSuggestions, type DomainSuggestionQuery } from '@automattic/api-core';
+import {
+	fetchDomainSuggestions,
+	fetchFreeDomainSuggestion,
+	type DomainSuggestionQuery,
+} from '@automattic/api-core';
 import { fetchDomains } from '@automattic/api-core';
 import { queryOptions } from '@tanstack/react-query';
 
@@ -9,10 +13,17 @@ export const domainsQuery = () =>
 	} );
 
 export const domainSuggestionsQuery = (
-	search: string,
-	query: Partial< DomainSuggestionQuery > = {}
+	query: string,
+	params: Partial< DomainSuggestionQuery > = {}
 ) =>
 	queryOptions( {
-		queryKey: [ 'domain-suggestions', search, query ],
-		queryFn: () => fetchDomainSuggestions( search, query ),
+		queryKey: [ 'domain-suggestions', query, params ],
+		queryFn: () => fetchDomainSuggestions( query, params ),
+		meta: { persist: false },
+	} );
+
+export const freeSuggestionQuery = ( query: string ) =>
+	queryOptions( {
+		queryKey: [ 'free-suggestion', query ],
+		queryFn: () => fetchFreeDomainSuggestion( query ),
 	} );
