+++ packages/api-queries/src/index.ts
@@ -1,5 +1,6 @@
 export * from './query-client';
 
+export * from './domain-availability';
 export * from './domain-dns-records';
 export * from './domain-dnssec';
 export * from './domain-forwarding';
