+++ packages/domain-search/src/page/context.ts
@@ -1,6 +1,9 @@
+import {
+	domainAvailabilityQuery,
+	domainSuggestionsQuery,
+	freeSuggestionQuery,
+} from '@automattic/api-queries';
 import { createContext, useCallback, useContext, useMemo, useState } from 'react';
-import { domainAvailabilityQuery } from '../queries/availability';
-import { domainSuggestionsQuery, freeSuggestionQuery } from '../queries/suggestions';
 import type { DomainSearchProps, DomainSearchContextType } from './types';
 
 const noop = () => {};
@@ -82,16 +85,21 @@ export const useDomainSearchContextValue = (
 			events: normalizedEvents,
 			config: normalizedConfig,
 			queries: {
-				domainSuggestions: ( query ) =>
-					domainSuggestionsQuery( query, {
+				domainSuggestions: ( query ) => ( {
+					...domainSuggestionsQuery( query, {
 						quantity: 30,
 						vendor: normalizedConfig.vendor,
 					} ),
+					enabled: false,
+				} ),
 				freeSuggestion: ( query ) => ( {
 					...freeSuggestionQuery( query ),
 					enabled: normalizedConfig.skippable,
 				} ),
-				domainAvailability: domainAvailabilityQuery,
+				domainAvailability: ( domainName ) => ( {
+					...domainAvailabilityQuery( domainName ),
+					enabled: false,
+				} ),
 			},
 			cart,
 			isFullCartOpen,
