--- packages/domain-search/src/queries/suggestions.ts
@@ -1,25 +1,0 @@
-import {
-	type DomainSuggestionQuery,
-	fetchDomainSuggestions,
-	fetchFreeDomainSuggestion,
-} from '@automattic/api-core';
-import { queryOptions } from '@tanstack/react-query';
-
-export const domainSuggestionsQuery = (
-	query: string,
-	params: Partial< DomainSuggestionQuery > = {}
-) =>
-	queryOptions( {
-		queryKey: [ 'domain-suggestions', query, params ],
-		queryFn: () => fetchDomainSuggestions( query, params ),
-		refetchOnWindowFocus: false,
-		refetchOnMount: false,
-		enabled: false,
-		meta: { persist: false },
-	} );
-
-export const freeSuggestionQuery = ( query: string ) =>
-	queryOptions( {
-		queryKey: [ 'free-suggestion', query ],
-		queryFn: () => fetchFreeDomainSuggestion( query ),
-	} );
